// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  USER: 'USER',
  PROFESSIONAL: 'PROFESSIONAL',
  ADMIN: 'ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const UserStatus = {
  ACTIVE: 'ACTIVE',
  BANNED: 'BANNED'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const LawRegion = {
  BANGLADESH: 'BANGLADESH',
  INTERNATIONAL: 'INTERNATIONAL'
} as const

export type LawRegion = (typeof LawRegion)[keyof typeof LawRegion]


export const LawCategory = {
  CYBERCRIME: 'CYBERCRIME',
  DATA_PROTECTION: 'DATA_PROTECTION',
  HARASSMENT: 'HARASSMENT',
  FINANCIAL: 'FINANCIAL',
  ELECTRONIC_EVIDENCE: 'ELECTRONIC_EVIDENCE',
  TELECOMMUNICATIONS: 'TELECOMMUNICATIONS',
  CHILD_PROTECTION: 'CHILD_PROTECTION',
  INTELLECTUAL_PROPERTY: 'INTELLECTUAL_PROPERTY',
  INTERNATIONAL_COOPERATION: 'INTERNATIONAL_COOPERATION',
  PLATFORM_GOVERNANCE: 'PLATFORM_GOVERNANCE'
} as const

export type LawCategory = (typeof LawCategory)[keyof typeof LawCategory]


export const LawStatus = {
  IN_FORCE: 'IN_FORCE',
  REFERENCE: 'REFERENCE',
  SUPERSEDED: 'SUPERSEDED',
  NOT_IN_FORCE: 'NOT_IN_FORCE'
} as const

export type LawStatus = (typeof LawStatus)[keyof typeof LawStatus]


export const ProfessionalRole = {
  POLICE: 'POLICE',
  LAWYER: 'LAWYER',
  PSYCHOLOGIST: 'PSYCHOLOGIST',
  HUMAN_RIGHTS: 'HUMAN_RIGHTS',
  ETHICAL_HACKER: 'ETHICAL_HACKER'
} as const

export type ProfessionalRole = (typeof ProfessionalRole)[keyof typeof ProfessionalRole]


export const HelpCategory = {
  LEGAL: 'LEGAL',
  MENTAL: 'MENTAL',
  EMERGENCY: 'EMERGENCY',
  ACCOUNT_SECURITY: 'ACCOUNT_SECURITY',
  ONLINE_HARASSMENT: 'ONLINE_HARASSMENT',
  FINANCIAL_FRAUD: 'FINANCIAL_FRAUD',
  IMAGE_ABUSE: 'IMAGE_ABUSE',
  CHILD_SAFETY: 'CHILD_SAFETY'
} as const

export type HelpCategory = (typeof HelpCategory)[keyof typeof HelpCategory]


export const HelpRequestStatus = {
  PENDING: 'PENDING',
  IN_PROGRESS: 'IN_PROGRESS',
  RESOLVED: 'RESOLVED',
  CLOSED: 'CLOSED'
} as const

export type HelpRequestStatus = (typeof HelpRequestStatus)[keyof typeof HelpRequestStatus]


export const MessageSender = {
  USER: 'USER',
  PROFESSIONAL: 'PROFESSIONAL'
} as const

export type MessageSender = (typeof MessageSender)[keyof typeof MessageSender]


export const SurveyQuestionType = {
  SHORT_TEXT: 'SHORT_TEXT',
  SINGLE_CHOICE: 'SINGLE_CHOICE',
  MULTIPLE_CHOICE: 'MULTIPLE_CHOICE',
  SCALE: 'SCALE'
} as const

export type SurveyQuestionType = (typeof SurveyQuestionType)[keyof typeof SurveyQuestionType]


export const SurveyResponseStatus = {
  COMPLETED: 'COMPLETED',
  DISCARDED: 'DISCARDED'
} as const

export type SurveyResponseStatus = (typeof SurveyResponseStatus)[keyof typeof SurveyResponseStatus]